job/requirements.txt
google-cloud-bigquery==3.27.0
requests==2.32.3
boto3==1.35.36
